requirements.txt
langchain==1.2.15
langchain-google-genai==4.2.2
google-generativeai==0.8.6
langchain-core==1.3.0
langgraph==1.1.9
python-dotenv==1.0.1
